Add zlib-ng-no-cmake feature
