+++ ci/test.bash
@@ -34,6 +34,11 @@ $CROSS run --target $TARGET_TRIPLE --manifest-path systest/Cargo.toml
 echo === zlib-ng build ===
 $CROSS test --target $TARGET_TRIPLE --no-default-features --features zlib-ng
 $CROSS run --target $TARGET_TRIPLE --manifest-path systest/Cargo.toml --no-default-features --features zlib-ng
+
+echo '=== zlib-ng-no-cmake build ==='
+$CROSS test --target "$TARGET_TRIPLE" --no-default-features --features zlib-ng-no-cmake
+$CROSS run --target "$TARGET_TRIPLE" --manifest-path systest/Cargo.toml --no-default-features --features zlib-ng-no-cmake
+
 echo === libz-ng-sys build ===
 mv Cargo-zng.toml Cargo.toml
 mv systest/Cargo-zng.toml systest/Cargo.toml
