--- src/compile_check/arm/arm_caps.c
@@ -1,4 +1,0 @@
-#include <sys/auxv.h>
-int main() {
-    return (getauxval(AT_HWCAP2) & HWCAP2_CRC32);
-}
