--- src/compile_check/arm/arm_hwcaps.c
@@ -1,5 +1,0 @@
-#include <sys/auxv.h>
-#include <asm/hwcap.h>
-int main() {
-    return (getauxval(AT_HWCAP2) & HWCAP2_CRC32);
-}
