--- src/compile_check/arm/arm_neon.c
@@ -1,4 +1,0 @@
-#include <sys/auxv.h>
-int main() {
-    return (getauxval(AT_HWCAP) & HWCAP_ARM_NEON);
-}
