--- src/compile_check/avx512/mask.c
@@ -1,3 +1,0 @@
-#include <immintrin.h>
-__mmask16 f(__mmask16 x) { return _knot_mask16(x); }
-int main(void) { return 0; }
