--- src/compile_check/flag.c
@@ -1,1 +1,0 @@
-int main() { return 0; }
